requirements.txt
groq
sentence-transformers
faiss-cpu
pypdf
gradio
python-dotenv
numpy>=2.1.0
aiohttp>=3.10.2